[TabTray] Move TabTray related classes to proper package
